Refactor backend code to use controller methods in routes
The backend code has been refactored to use controller methods instead of inline functions in route definitions. This has significantly reduced the lines of code in route files and improved their readability. This refactoring includes changes in manufacturing, customer, order, and items routes as well as corresponding controller files.
