+++ apps/backend/package.json
@@ -4,10 +4,10 @@
   "main": "index.js",
   "author": "Bjorn Francke <bjorn@francke.dev>",
   "devDependencies": {
+    "@types/cors": "^2.8.17",
     "@types/express": "^4.17.21",
     "nodemon": "^3.1.0",
     "ts-node-dev": "^2.0.0",
-    "@types/cors": "^2.8.17",
     "typescript": "^5.4.3"
   },
   "scripts": {
@@ -20,6 +20,7 @@
     "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.1",
+    "express-async-handler": "^1.2.0",
     "mongodb": "^6.5.0",
     "mongoose": "^8.2.3"
   }
