+++ apps/backend/src/controllers/index.ts
@@ -1,0 +1,5 @@
+export * from "./itemsController"
+export * from "./bomController"
+export * from "./orderController"
+export * from "./customerController"
+export * from "./manufacturingController"
