+++ apps/backend/src/routes/items/index.ts
@@ -1,66 +1,26 @@
 import express from "express";
-import {Item, listAllItems} from "../../items";
-import {Error} from "mongoose";
-import {Bom} from "../../bom";
+import {Item} from "../../items";
+import {createItem, deleteItem, getItems, getItemsById, setItemStock} from "../../controllers";
 
 
 const itemRouter = express.Router()
 
-
 itemRouter.route('/findAll')
-    .get(async(req, res) => {
-    try {
-        const response = await listAllItems()
-        res.json(response)
-    } catch {
-        res.status(500).send("Internal Server Error");
-    }
-    })
+    .get(getItems)
 
 itemRouter.route('/delete/:itemId')
-    .delete(async (req, res) => {
-        try {
-            const isItemUsedInBom = await Bom.findOne({ "components.id": req.params.itemId });
-            if (isItemUsedInBom) {
-                return res.status(403).send("Item is used in a BOM and cannot be deleted");
-            }
-            const deletedItem = await Item.findByIdAndDelete(req.params.itemId);
-            res.json(deletedItem)
-        } catch {
-            res.status(500).send("Internal Server Error");
-        }
-    })
+    .delete(deleteItem)
 
 itemRouter.route('/create')
-    .post(async (req, res) => {
-        try {
-            const newItem = new Item(
-                {
-                    name: req.body.name,
-                    description: req.body.description,
-                    stock: req.body.stock
-                }
-            );
-            const savedItem = await newItem.save();
-            res.json(savedItem);
-        } catch {
-            res.status(500).send("Internal Server Error");
-        }
-    })
+    .post(createItem)
 
 
 itemRouter.route('/findById/:itemId')
-    .get(async (req, res) => {
-    try {
-        const itemData = await Item.findById(req.params.itemId)
-        if (!itemData) {
-            res.status(404).send("Item was not found")
-        }
-        res.json(itemData)
-    } catch {
-        res.status(500).send("Internal Server Error")
-    }
-})
+    .get(getItemsById)
+
+itemRouter.route('/setStock/:itemId/:newStock')
+    .patch(setItemStock)
+
 
 itemRouter.route('/getNameById/:itemId')
     .get(async (req, res) => {
@@ -92,24 +52,9 @@ itemRouter.route('/update/:itemId')
         }
     })
 
-itemRouter.route('/setStock/:itemId/:newStock')
-    .patch(async (req, res) => {
-        try{
-            const id = req.params.itemId
-            const newStock: number = Number(req.params.newStock)
-            const item = await Item.findById(id);
-            if (!item) {
-                res.status(404).send("Item was not found")
-                return
-            }
-            item.stock = newStock;
 
-            const updatedItem = await item.save();
-            res.send(updatedItem);
 
-        } catch {
-            res.status(500).send("Internal Server Error");
-        }
-    })
+
+
 
 export default itemRouter
